feat(chess): add basic rules to chess
I write the basic rules of eight kinds of chess pieces, and added some methods to judge the attributes and rationality in Chess class.

I also added the getPossibleMovesOnOneDirection method to this class, which is used for the chess judgment of queen, bishop and rook.

I changed the original color interface to enum type, because I think enum here will have better operability than the interface, and also simplify some classes.