feat(chess): add specific rules to chess and modify some irrational codes
I achieve the special moves of several pieces in chess, such as castling, capture pawn en passant, and the oblique eating of pawn.

I also modified some inappropriate codes. For example, in the direction enumeration class, I adjusted the values encapsulated in each Direction object, so that the direction and the corresponding value can be directly mapped.

I adjusted the names of some methods and properties to make them sound concise and easy to understand.