feat(chess, game): add moveChess method and create mementos
I write move method of playing chess. This method determines the player's chess strategy through ChessStrategy class, which stores the start Position of the chess and the end Point of it.

In addition, the recall function of the game is also completed. I combine the Memento Pattern with the Prototype Pattern. The information in current game will be managed through the GameCaretaker class, and the info will be returned to the previous game state through the undo() method in the Game class.