feat(chess, game): add check and mate judge, complete System class
I achieve the judgment of winning, including whether the king can avoid the attack and whether no chess pieces can eat the chess pieces that are attacking the king.

During the test, I found and fixed some bugs in the special rules of chess.The GameSystem class is also implemented. This class is used to wrap archive information for easy read / write operations and add / delete operations.